Adicionado o parâmetro "deleteAll" no teste do banco de dados como flag para apagar todos os registros

+++ app/src/main/java/com/amiltonsoft/amiltonjunior/pessoalandroid/DB.java
@@ -104,7 +104,7 @@ public final class DB {
     }
 
     // Método que testa o funcionamento do banco de dados
-    public void testDB() {
+    public void testDB(boolean deleteAll) {
         // Inserir dados
         this.insert("Amilton Junior", 24, "M");
         this.insert("Deise Carolina", 25, "F");
@@ -132,7 +132,8 @@ public final class DB {
         }
 
         // Apaga todos os dados do banco de dados
-        this.deleteAll();
+        if (deleteAll)
+            this.deleteAll();
     }
 
     /****************************/
